spark-app/requirements.txt
requests
py4j
kafka-python-ng
pyspark
six
